requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.7
filelock==3.12.4
idna==3.4
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.3
platformdirs==3.10.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
splitwise==3.0.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.5
